pyproject.toml
[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[project]
name = "programstart-workflow"
version = "0.1.0"
description = "Workflow tooling, validation, and dashboard automation for PROGRAMSTART."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "filelock>=3.0",
  "pydantic>=2.0",
]
keywords = ["planning", "workflow", "automation", "PROGRAMBUILD", "stage-gate"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Office/Business",
  "Topic :: Software Development :: Quality Assurance",
]

[project.optional-dependencies]
dev = [
  "bandit>=1.8.3",
  "check-jsonschema>=0.37.1",
  "coverage>=7.13.5",
  "detect-secrets>=1.5.0",
  "mkdocs>=1.6.1",
  "mkdocs-material>=9.7.6",
  "nox>=2025.10.16",
  "pip-audit>=2.9.0",
  "pillow>=11.2.1",
  "playwright>=1.54.0",
  "pre-commit>=4.5.0",
  "pyright>=1.1.403",
  "pytest>=8.4.1",
  "pytest-cov>=7.0.0",
  "ruff>=0.12.10",
  "uv>=0.8.17",
  "yamllint>=1.37.0",
  "hypothesis>=6.100.0",
  # mypy removed: pyright is the active type checker (pre-commit, nox, CI).
  # Running both creates redundant gates with conflicting error sets.
]
vector = [
  "chromadb>=0.4.0",
]
rag = [
  "chromadb>=0.4.0",
  "litellm>=1.40.0",
  "instructor>=1.3.0",
]

[project.scripts]
programstart = "scripts.programstart_cli:main"
programstart-retrieval = "scripts.programstart_retrieval:main"
programstart-context = "scripts.programstart_context:main"
programstart-clean = "scripts.programstart_clean:main"
programstart-status = "scripts.programstart_status:main"
programstart-validate = "scripts.programstart_validate:main"
programstart-bootstrap = "scripts.programstart_bootstrap:main"
programstart-workflow = "scripts.programstart_workflow_state:main"
programstart-drift = "scripts.programstart_drift_check:main"
programstart-guide = "scripts.programstart_step_guide:main"
programstart-init = "scripts.programstart_init:main"
programstart-create = "scripts.programstart_create:main"
programstart-attach = "scripts.programstart_attach:main"
programstart-recommend = "scripts.programstart_recommend:main"
programstart-impact = "scripts.programstart_impact:main"
programstart-prompt-eval = "scripts.programstart_prompt_eval:main"
programstart-doctor = "scripts.programstart_doctor:main"
programstart-prompt-build = "scripts.programstart_prompt_build:main"

[tool.setuptools]
packages = ["scripts"]

[tool.ruff]
target-version = "py312"
line-length = 130
extend-exclude = [
  "BACKUPS",
  "build",
  "dist",
  ".tmp_bootstrap_real",
  ".tmp_bootstrap_uvverify",
  ".tmp_bootstrap_verify",
  ".tmp_ci_bootstrap",
  ".tmp_dist_smoke",
  "site",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"scripts/programstart_dashboard_browser_smoke.py" = ["E501"]
"scripts/programstart_dashboard_smoke.py" = ["E501"]
"scripts/programstart_serve.py" = ["E501"]
"tests/*.py" = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["programstart_common"]

[tool.pyright]
include = ["scripts", "tests", "noxfile.py"]
exclude = [
  "BACKUPS",
  "build",
  "dist",
  ".tmp_bootstrap_real",
  ".tmp_bootstrap_verify",
  ".tmp_ci_bootstrap",
  ".tmp_dist_smoke",
  "site",
]
extraPaths = ["scripts"]
typeCheckingMode = "basic"
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"
reportMissingTypeStubs = false

[tool.pytest.ini_options]
addopts = "-ra --strict-markers"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = [
  "scripts.programstart_cli",
  "scripts.programstart_backup",
  "scripts.programstart_context",
  "scripts.programstart_clean",
  "scripts.programstart_closeout",
  "scripts.programstart_command_registry",
  "scripts.programstart_common",
  "scripts.programstart_validate",
  "scripts.programstart_bootstrap",
  "scripts.programstart_create",
  "scripts.programstart_create_core",
  "scripts.programstart_create_output",
  "scripts.programstart_status",
  "scripts.programstart_workflow_state",
  "scripts.programstart_drift_check",
  "scripts.programstart_step_guide",
  "scripts.programstart_log",
  "scripts.programstart_retrieval",
  "scripts.programstart_models",
  "scripts.programstart_attach",
  "scripts.programstart_init",
  "scripts.programstart_impact",
  "scripts.programstart_recommend",
  "scripts.programstart_checklist_progress",
  "scripts.programstart_refresh_integrity",
  "scripts.programstart_dashboard",
  "scripts.programstart_serve",
  "scripts.programstart_starter_scaffold",
  "scripts.programstart_markdown_parsers",
  "scripts.programstart_prompt_eval",
  "scripts.programstart_health_probe",
  "scripts.programstart_doctor",
  "scripts.programstart_prompt_build",
  "scripts.lint_prompts",
  "scripts.install_hooks",
  "scripts.programstart_research_delta",
  "scripts.check_commit_msg",
  "scripts.programstart_repo_clean_check",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 90
